backend_v2/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
alembic==1.13.1
cryptography==41.0.7
tenacity==8.2.3
